- Fixed typo in glcontext.cpp that caused compile errors - Fixed pow calls for building with free MS C++ compiler

+++ trunk/celestia/src/celengine/render.cpp
@@ -233,7 +233,7 @@ static void PenumbraFunctionEval(float u, float v, float w,
     u = (u + 1.0f) * 0.5f;
 
     // Using the cube root produces a good visual result
-    unsigned char pixVal = (unsigned char) (::pow(u, 0.33) * 255.99);
+    unsigned char pixVal = (unsigned char) (::pow((double) u, 0.33) * 255.99);
 
     pixel[0] = pixVal;
 }
@@ -3682,7 +3682,6 @@ static void renderSphere_GLSL(const RenderInfo& ri,
 
     setLightParameters_GLSL(*prog, shadprop, ls,
                             ri.color, ri.specularColor);
-
 
     prog->shininess = ri.specularPower;
     prog->ambientColor = Vec3f(ri.ambientColor.red(), ri.ambientColor.green(),
